parser/requirements.txt
pandas
pyarrow
pymongo
google-cloud-storage
gcsfs